fix: 와일드카드 제거 #17
